requirements.txt
python-telegram-bot==21.3
gspread==6.1.2
google-auth==2.35.0
